Remove nearly-colinear vertices
This check will now cause vertices to also be removed if they are nearly colinear, with a deviation of up to 2.5 micrometres.

This fix is based on this commit: https://github.com/Ultimaker/CuraEngine/pull/1214/commits/1afa4507c180caea34dd0e08d94546749842760c which fixes the issue too. I tried to limit the impact on the global polygon shape as little as possible, but this fix will have some impact.
Let's see if we want to merge this, and if so where to.

Contributes to issue CURA-7393.